Search toolbar optimizing
1.

The searchOperators are totally useless for checkbox > So avoid them for this case.

The clearSearch is useless for select since there is no multiselect  in search toolbar > Avoid it..

Also, it saves some width on these columns.

2.

remove dirty inline styles.

remove <table> search that is inappropriate and replace all with spans but search operator have now no container

change css accordingly to js change.