fix usage of wrong column headers in case of usage hidden: "true"
One posts from time to time the same problem with wrong column headers
in case of usage `hidden: "true"` instead of `hidden: true`. jqGrid hide
the column of data in the case, but don't hide the column headers. As
the result one have grid with vary strange layout. The fix uses only
boolean values of `hidden` property and the usage of `hidden: "true"`
hide nothing.