Oleg change - removing unneeded &nbsp; from data field of View and Edit dialogs
The existence of unneeded `&nbsp;` at the beginning of the data field
can break a little the format of data displayed, especially in case of
long multiline data which are inline elements (like `<span>` if the data
field of the View dialog). See [the old
answer](http://stackoverflow.com/a/8171924/315935) to more details.