Oleg change - add skipUsageOfJsonMap option used mostly internally, change reloadGrid in case of usage with loadonce:true
The new option `skipUsageOfJsonMap` is included to fix the compatibility
issue in case of usage `jsonmap` (or `xmlmap`) property of `colModel`
with remote `datatype` (`"json"` or `"xml"`) together with `loadonce:
true`. The option `skipUsageOfJsonMap:true` will be set by jqGrid
directly after automatically changing of `datatype` from the original
`"json"` or `"xml"` value to `"local"`. It forces the usage of `jsonmap`
(or `xmlmap`) *only with remote data*. The original value of `datatype`
will be saved in internal `dataTypeOrg` option. It allows to modify the
behavior of `reloadGrid` to force reloading of *remote* data from the
server.